requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
streamlit==1.43.0
openai==1.58.1
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
